Remove legacy code only keep shamela flow
