+++ package.json
@@ -9,13 +9,13 @@
         "ansis": "^4.1.0",
         "baburchi": "^1.5.0",
         "baheth-sdk": "^1.2.2",
-        "bitaboom": "^1.4.1",
+        "bitaboom": "^1.5.0",
         "cli-progress": "^3.12.0",
         "cli-welcome": "^3.0.1",
         "conf": "^14.0.0",
         "kokokor": "^2.0.4",
         "linkedom": "^0.18.12",
-        "pino": "^9.9.4",
+        "pino": "^9.9.5",
         "pino-pretty": "^13.1.1",
         "rabbito": "^1.0.0",
         "remove-markdown": "^0.6.2",
@@ -33,10 +33,10 @@
         "eslint-config-prettier": "^10.1.8",
         "eslint-plugin-perfectionist": "^4.15.0",
         "eslint-plugin-prettier": "^5.5.4",
-        "globals": "^16.3.0",
+        "globals": "^16.4.0",
         "prettier": "^3.6.2",
-        "semantic-release": "^24.2.7",
-        "typescript-eslint": "^8.42.0"
+        "semantic-release": "^24.2.8",
+        "typescript-eslint": "^8.43.0"
     },
     "engines": {
         "bun": ">=1.2.16"
