--- src/actions/translate/prompt.ts
@@ -1,115 +1,0 @@
-import { parseArgs } from 'node:util';
-
-import { getNumericInput } from '../../utils/io.js';
-
-export const promptTranslateInputs = async () => {
-    const { values } = parseArgs({
-        options: {
-            collection: {
-                type: 'string',
-            },
-            diff: {
-                type: 'string',
-            },
-            discrete: {
-                type: 'boolean',
-            },
-            explains: {
-                type: 'string',
-            },
-            fix: {
-                type: 'boolean',
-            },
-            footnotes: {
-                type: 'boolean',
-            },
-            interactive: {
-                type: 'boolean',
-            },
-            pages: {
-                type: 'string',
-            },
-            preview: {
-                type: 'boolean',
-            },
-            refresh: {
-                type: 'boolean',
-            },
-            translate: {
-                type: 'boolean',
-            },
-            translator: {
-                type: 'string',
-            },
-            url: {
-                type: 'string',
-            },
-            walk: {
-                type: 'boolean',
-            },
-        },
-        strict: true,
-    });
-
-    const diff = values.diff
-        ? Number(values.diff.startsWith('n') ? '-' + values.diff.slice(1) : values.diff)
-        : undefined;
-
-    const collectionId =
-        values.collection ||
-        (await getNumericInput('Enter collection ID to translate:', 'Please enter a valid collection ID'));
-
-    if (diff) {
-        return {
-            autoFix: false,
-            collectionId,
-            diff,
-            fromPage: 1,
-            isPreview: true,
-            refresh: false,
-            removeFootnotes: false,
-            toPage: 1,
-            translatorId: '1',
-        };
-    }
-
-    const translatorId =
-        values.translator || (await getNumericInput('Enter translator ID:', 'Please enter a valid translator ID'));
-
-    const fromPage =
-        values.pages?.split('-')[0] ||
-        (await getNumericInput('Enter page to start from:', 'Please enter a valid page number', '1'));
-
-    const toPage =
-        values.pages?.split('-').at(-1) ||
-        (await getNumericInput(
-            'Enter page to end:',
-            'Please enter a valid page number',
-            Number.MAX_SAFE_INTEGER.toString(),
-        ));
-
-    let strategy = '';
-
-    if (values.discrete) {
-        strategy = 'discrete';
-    } else if (values.walk) {
-        strategy = 'walk';
-    } else if (values.interactive) {
-        strategy = 'interactive';
-    }
-
-    return {
-        autoFix: Boolean(values.fix),
-        collectionId,
-        diff,
-        explains: values.explains,
-        fromPage: Number(fromPage),
-        isPreview: Boolean(values.preview),
-        refresh: Boolean(values.refresh),
-        removeFootnotes: !values.footnotes,
-        strategy,
-        toPage: Number(toPage),
-        translatorId,
-        url: values.url,
-    };
-};
