--- src/actions/translate/types.ts
@@ -1,5 +1,0 @@
-import type { Page } from '../../api/maktabah.js';
-
-export type PageRange = Page & {
-    end?: number;
-};
