+++ src/actions/translator.ts
@@ -1,0 +1,1 @@
+export const translateUsingAI = () => {};
